LayoutTests/js/dfg-mispredict-variable-but-prove-int-expected.txt

Tests that a variable predicted to be either int or double but proven to be an int does confuse the Branch logic.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS foo(0, 1, 0) is 0.5
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS foo(i, i + 1, i + 2) is 1
PASS successfullyParsed is true

TEST COMPLETE
